(wip) Serve an OIDC authorization_endpoint and end_session_endpoint

+++ app/controllers/logout_controller.rb
@@ -2,8 +2,20 @@ require "services"
 
 class LogoutController < ApplicationController
   def show
-    session.delete(:sub)
+    # TODO: delete our keycloak cookies
 
-    redirect_to "#{Services.discover.end_session_endpoint}?post_logout_redirect_uri=https://www.gov.uk"
+    if session[:sub]
+      Services.keycloak.users.logout(session[:sub])
+      session.delete(:sub)
+    end
+
+    # TODO: we should restrict this URI to avoid an open redirect
+    redirect_to logout_params.fetch(:post_logout_redirect_uri, "https://www.gov.uk")
+  end
+
+private
+
+  def logout_params
+    params.permit(:post_logout_redirect_uri)
   end
 end
